Changes for 2021-03-02

+++ tools/build/lib/nhsx/tasks/download.rb
@@ -52,8 +52,9 @@ namespace :download do
   end
 
   desc "Download the code build artifacts"
-  task :"codebuild:dev" => [:"login:dev"] do
+  task :"codebuild:dev" => [:"login:dev", :"clean:wipe"] do
     include NHSx::Report
+
     # get build info
     job_id = $configuration.job_id
     build_info = build_info([job_id])
@@ -63,12 +64,17 @@ namespace :download do
     object_name = build_info.first["artifacts"]["location"]
     object_name = object_name.sub("arn:aws:s3:::", "")
     zip_file_path = File.join(downloads_out_dir, "#{object_name}.zip")
-    run_command("Download the build artifacts of #{job_id}", NHSx::AWS::Commandlines.download_from_s3(object_name, zip_file_path), $configuration)
-
-    # unzip to base dir
-    run_command("Unzip archive", "unzip #{zip_file_path} -d #{$configuration.base}", $configuration)
-
-    # remove downloaded s3 zip file
-    File.delete(zip_file_path)
+    if object_name == "dev-build-artifacts-archive/ci-app-system"
+      puts "No upload artifacts present for build #{job_id}"
+    else
+      run_command("Download the build artifacts of #{job_id}", NHSx::AWS::Commandlines.download_from_s3(object_name, zip_file_path), $configuration)
+      # unzip to base dir
+      run_command("Unzip archive", "unzip #{zip_file_path} -d #{$configuration.base}", $configuration)
+      # remove downloaded s3 zip file
+      File.delete(zip_file_path)
+    end
+    include NHSx::Queue #redefines build_info - needs to be cleaned up
+    bi = NHSx::Queue::CodeBuildInfo.new(build_info.first)
+    pipe_logs(bi)
   end
 end
